fix(avatar): body base layer, user_id BIGINT, API debug mode
- Add fixed body layer: assets/avatars/base/body.svg (fallback to _placeholder)
- Layer order: bg -> body -> bottom -> shoes -> top -> accessory -> hair -> frame
- Fix user_id INT->BIGINT in knd_user_avatar and knd_user_avatar_inventory (FK to users)
- API state/equip: return error detail when ?debug=1 for troubleshooting

Made-with: Cursor

+++ sql/knd_user_avatar.sql
@@ -2,7 +2,7 @@
 -- Run after knd_avatar_items.sql, requires users table
 
 CREATE TABLE IF NOT EXISTS `knd_user_avatar` (
-  `user_id` INT NOT NULL,
+  `user_id` BIGINT NOT NULL,
   `hair_item_id` INT NULL,
   `top_item_id` INT NULL,
   `bottom_item_id` INT NULL,
